[backend] Created custom router
This router features only one new thing: it allows a viewset to define a function `get_lookup_regex`
which will allow a viewset to precisely define the exact URL regex.

Also switches the URLpattern to fetch a single map from zoom/x/z to x/z/zoom.
This will allow for the zoom parameter to become optional in future.